+++ druid-derive/tests/ui/with-union.rs
@@ -1,0 +1,9 @@
+use druid::*;
+
+#[derive(Lens)]
+union Foobar {
+    foo: i32,
+    bar: f64,
+}
+
+fn main() {}